Refactor: rename expense app name to fyle_expense
